scheduler/requirements.txt
aiohttp==3.8.1
aiokafka==0.7.2
python-dotenv==0.19.2
psycopg2-binary==2.9.1
sqlalchemy==1.4.22
SQLAlchemy-Utils==0.37.8
alembic==1.7.5
fastapi==0.68.0
uvicorn[standard]==0.15.0
